+++ vscode/src/gh-copilot/azureQuantumTools.ts
